+++ src/Markdig/Parsers/ParagraphBlockParser.cs
@@ -12,6 +12,8 @@ namespace Markdig.Parsers
     /// <seealso cref="BlockParser" />
     public class ParagraphBlockParser : BlockParser
     {
+        public bool ParseSetexHeadings { get; set; } = true;
+
         public override BlockState TryOpen(BlockProcessor processor)
         {
             if (processor.IsBlankLine)
@@ -35,7 +37,7 @@ namespace Markdig.Parsers
                 return BlockState.BreakDiscard;
             }
 
-            if (!processor.IsCodeIndent && !(block.Parent is QuoteBlock))
+            if (ParseSetexHeadings && !processor.IsCodeIndent && !(block.Parent is QuoteBlock))
             {
                 return TryParseSetexHeading(processor, block);
             }
@@ -46,8 +48,7 @@ namespace Markdig.Parsers
 
         public override bool Close(BlockProcessor processor, Block block)
         {
-            var paragraph = block as ParagraphBlock;
-            if (paragraph != null)
+            if (block is ParagraphBlock paragraph)
             {
                 TryMatchLinkReferenceDefinition(ref paragraph.Lines, processor);